refactor arg checks to own module
